LibCrypto: Add changable CRC32 polynomial
