+++ Userland/Libraries/LibCrypto/CMakeLists.txt
@@ -19,7 +19,6 @@ set(SOURCES
     BigInt/UnsignedBigInteger.cpp
     Checksum/Adler32.cpp
     Checksum/cksum.cpp
-    Checksum/CRC32.cpp
     Cipher/AES.cpp
     Cipher/ChaCha20.cpp
     Curves/Curve25519.cpp
